+++ src/http/header/accept_charset.rs
@@ -22,11 +22,11 @@ crate::http::header::common_header! {
     /// # Examples
     /// ```
     /// use actix_web::HttpResponse;
-    /// use actix_web::http::header::{AcceptCharset, Charset, qitem};
+    /// use actix_web::http::header::{AcceptCharset, Charset, QualityItem};
     ///
     /// let mut builder = HttpResponse::Ok();
     /// builder.insert_header(
-    ///     AcceptCharset(vec![qitem(Charset::Us_Ascii)])
+    ///     AcceptCharset(vec![QualityItem::max(Charset::Us_Ascii)])
     /// );
     /// ```
     ///
@@ -45,11 +45,11 @@ crate::http::header::common_header! {
     ///
     /// ```
     /// use actix_web::HttpResponse;
-    /// use actix_web::http::header::{AcceptCharset, Charset, qitem};
+    /// use actix_web::http::header::{AcceptCharset, Charset, QualityItem};
     ///
     /// let mut builder = HttpResponse::Ok();
     /// builder.insert_header(
-    ///     AcceptCharset(vec![qitem(Charset::Ext("utf-8".to_owned()))])
+    ///     AcceptCharset(vec![QualityItem::max(Charset::Ext("utf-8".to_owned()))])
     /// );
     /// ```
     (AcceptCharset, ACCEPT_CHARSET) => (QualityItem<Charset>)+